Consolidate redux calls to PrivacyDeclarationManager
The redux calls had originally been isolated to a parent component to facilitate reuse across redux environments, but we do not need that anymore
